user authentication added to create post

+++ pages/api/products/[id].js
@@ -1,6 +1,5 @@
 import dbConnect from '../../../util/mongo'
 import Product from '../../../models/Product'
-import Order from '../../../models/Order'
 
 export default async function handler(req, res) {
   const {
@@ -9,6 +8,7 @@ export default async function handler(req, res) {
     cookies,
   } = req
 
+  const token = cookies.token
   await dbConnect()
 
   if (method === 'GET') {
@@ -21,6 +21,9 @@ export default async function handler(req, res) {
   }
 
   if (method === 'PUT') {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json('Not Authenticated!')
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -32,6 +35,9 @@ export default async function handler(req, res) {
   }
 
   if (method === 'DELETE') {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json('Not Authenticated!')
+    }
     try {
       await Product.findByIdAndDelete(id)
       res.status(201).json('The product has been deleted')
